test refactoring: move MzVersion to a separate file
