+++ misc/python/materialize/checks/all_checks/pg_cdc.py
@@ -13,7 +13,7 @@ from typing import Any
 
 from materialize.checks.actions import Testdrive
 from materialize.checks.checks import Check, disabled
-from materialize.util import MzVersion
+from materialize.mz_version import MzVersion
 
 
 class PgCdcBase:
