+++ misc/python/materialize/checks/all_checks/temporal_types.py
@@ -11,7 +11,7 @@ from textwrap import dedent
 from materialize.checks.actions import Testdrive
 from materialize.checks.checks import Check
 from materialize.checks.executors import Executor
-from materialize.util import MzVersion
+from materialize.mz_version import MzVersion
 
 
 class TemporalTypes(Check):
