+++ misc/python/materialize/checks/mzcompose_actions.py
@@ -12,9 +12,9 @@ from typing import TYPE_CHECKING, Any
 
 from materialize.checks.actions import Action
 from materialize.checks.executors import Executor
+from materialize.mz_version import MzVersion
 from materialize.mzcompose.services.clusterd import Clusterd
 from materialize.mzcompose.services.materialized import Materialized
-from materialize.util import MzVersion
 
 if TYPE_CHECKING:
     from materialize.checks.scenarios import Scenario
