+++ test/legacy-upgrade/mzcompose.py
@@ -13,6 +13,7 @@ operational after an upgrade.
 
 import random
 
+from materialize.mz_version import MzVersion
 from materialize.mzcompose.composition import Composition, WorkflowArgumentParser
 from materialize.mzcompose.services.cockroach import Cockroach
 from materialize.mzcompose.services.kafka import Kafka
@@ -22,7 +23,6 @@ from materialize.mzcompose.services.schema_registry import SchemaRegistry
 from materialize.mzcompose.services.test_certs import TestCerts
 from materialize.mzcompose.services.testdrive import Testdrive
 from materialize.mzcompose.services.zookeeper import Zookeeper
-from materialize.util import MzVersion
 from materialize.version_list import VersionsFromDocs
 
 version_list = VersionsFromDocs()
